pyproject.toml
[project]
name = "touhou-character-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "contourpy>=1.3.3",
    "cycler==0.12.1",
    "filelock==3.19.1",
    "fonttools==4.60.1",
    "fsspec==2025.10.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "kiwisolver>=1.4.9",
    "markupsafe==3.0.3",
    "matplotlib>=3.10.8",
    "mpmath==1.3.0",
    "networkx==3.2.1",
    "numpy>=2.4.1",
    "packaging==25.0",
    "pillow==11.3.0",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "streamlit>=1.53.0",
    "sympy==1.14.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "typing-extensions==4.15.0",
    "zipp==3.23.0",
]
